Refactor viewer into modules
